Article Contents + Quiz
These are the 8 articles that will be on display in the website's "Education" tab. The quiz is also included. Will be working on styling and integration (i.e. adding a comprehension checker feature in each article, allowing for seamless navigation between each article, and for the final quiz to be taken whenever).